move legacy scheme to new module
